feat: add `Dockerfile` template
# Motivation

More and more applications are being run natively in containers, for a
variety of reasons. This commit is the initial attempt to provide
containerized support for `SolrWayback`.

# Usage

To get going, simply run the following command:

```bash
docker build . --tag solrwayback
docker run --publish 8080:8080 --publish 8983:8983  --volume \
 <path/to/WARCs>:/host_dir --tty --interactive solrwayback bash
```
where `<path/to/WARCs>` only contains `WARC` files and directories.

For more details, please refer to the comments in the `Dockerfile`.

# Implementation

The `Dockerfile` was created by following the instructions in the
top-level readme. In addition to this, a few verification steps were
added to ensure that the container works as expected.

A simple test that uses `docker` to build the container has been added.

# Drawbacks

## No proper indexing test

The added test does not verify that indexing works as expected. This
should be done by adding some `WARC` files as test data, but this is
outside the scope of this commit.

## No automatic update to latest release

Whenever someone releases a new version of `SolrWayback`, there is no
reminder or automatic failure if the relevant people forget to update
the `Dockerfile`.

# Future work

## Index preservation

The `Dockerfile` does currently not preserve the created index, so it
needs to be manually copied out of the container in order to preserve
it. With the latest released `SolrWayback` bundle, the index can be
found here:
`unpacked-bundle/solrwayback_package_4.4.2/solr-7.7.3/server/solr/
configsets/netarchivebuilder/netarchivebuilder_data/index/`

## Custom configuration of `properties`

There is currently no way of using your own `solrwayback.properties` and
`solrwaybackweb.properties`, which is essential for using the correct
branding.