Set explicit encoding UTF-8 for indexing
The warc-indexer used for indexing WARCs to Solr does not specify UTF-8 as the
output character encoding itself. While UTF-8 is default on many systems, it is
still an assumption and if e.g. ASCII is the default, all non-ASCII characters
will be replaced with an "unknown character" representation (typically a question
mark in a black rhombus). This was triggered by Trym Brennes while running the
bundle under Docker.

Setting file.encoding to UTF-8 is a workaround until the warc-indexer is changed
to always output UTF-8. As the output is used for Solr (or Elasticsearch) which
both supports UTF-8, this is the only encoding that makes sense.