Refactor SolrGenericStreaming to focus on creating streams or iterators for SolrDocuments
In order to implement #329 it will be necessary to use different providers of SolrDocuments as well as a merger mechanism. As the current code only uses static methods SolrGenericStreaming as well as the stream of SolrDocuments that it delivers, it makes sense to hide the SolrGenericStreaming instance and replace it with methods for generating streams and iterators of SolrDocuments.